fixed bug in create_cutout_multithread

+++ pyaedt/edb.py
@@ -832,7 +832,8 @@ class Edb(object):
         self.logger.info("EDB file release time: {0:.2f}ms".format(elapsed_time * 1000.0))
         self._clean_variables()
         timeout = 4
-        time.sleep(2)
+        # TODO check if we can remove this sleep
+        # time.sleep(2)
         while gc.collect() != 0 and timeout > 0:
             time.sleep(1)
             timeout -= 1
@@ -1114,6 +1115,7 @@ class Edb(object):
         number_of_threads=4,
         custom_extent=None,
         output_aedb_path=None,
+        remove_single_pin_components=False,
     ):
         """Create a cutout using an approach entirely based on pyaedt.
         It does in sequence:
@@ -1144,6 +1146,8 @@ class Edb(object):
             Edb PolygonData object. In this case, both signal_list and reference_list will be cut.
         output_aedb_path : str, optional
             Full path and name for the new AEDB file. If None, then current aedb will be cutout.
+        remove_single_pin_components : bool, optional
+            Remove all Single Pin RLC after the cutout is completed. Default is `False`.
 
         Returns
         -------
@@ -1230,6 +1234,9 @@ class Edb(object):
         def intersect(poly1, poly2):
             return list(poly1.Intersect(poly2))
 
+        def subtract(poly, voids):
+            return poly.Subtract(convert_py_list_to_net_list(poly), convert_py_list_to_net_list(voids))
+
         def clean_prim(prim_1):
             net = prim_1.net_name
             if net in reference_list:
@@ -1240,17 +1247,35 @@ class Edb(object):
                 elif int_data != 2:
                     list_poly = intersect(_poly, get_polygon_data(prim_1))
                     if list_poly:
-                        list_void = []
                         voids = prim_1.voids
-                        if voids:
-                            for void in voids:
-                                void_pdata = get_polygon_data(void)
-                                int_data = _poly.GetIntersectionType(void_pdata)
-                                if int_data != 2 and int_data != 0:
-                                    list_void.extend(intersect(_poly, void_pdata))
-                                elif int_data == 2:
-                                    list_void.append(void_pdata)
-                        poly_to_create.append([list_poly, prim_1.layer_name, net, list_void])
+                        for p in list_poly:
+                            if p.IsNull():
+                                continue
+                            list_void = []
+                            void_to_subtract = []
+                            if voids:
+                                for void in voids:
+                                    void_pdata = get_polygon_data(void)
+                                    int_data = p.GetIntersectionType(void_pdata)
+                                    if int_data > 2:
+                                        void_to_subtract.append(void_pdata)
+                                    elif int_data == 2:
+                                        list_void.append(void_pdata)
+                                if void_to_subtract:
+                                    polys_cleans = subtract(p, void_to_subtract)
+                                    for polys_clean in polys_cleans:
+                                        if not polys_clean.IsNull():
+                                            void_to_append = [
+                                                v for v in list_void if polys_clean.GetIntersectionType(v) == 2
+                                            ]
+                                            poly_to_create.append(
+                                                [[polys_clean], prim_1.layer_name, net, void_to_append]
+                                            )
+                                else:
+                                    poly_to_create.append([[p], prim_1.layer_name, net, list_void])
+                            else:
+                                poly_to_create.append([[p], prim_1.layer_name, net, list_void])
+
                     prims_to_delete.append(prim_1)
 
         def pins_clean(pinst):
@@ -1280,12 +1305,14 @@ class Edb(object):
         self.logger.info_timer("Primitives cleanup completed")
         self.logger.reset_timer()
 
-        self.core_components.delete_single_pin_rlc()
         i = 0
         for comp, val in self.core_components.components.items():
             if val.numpins == 0:
                 val.edbcomponent.Delete()
                 i += 1
+        if remove_single_pin_components:
+            self.core_components.delete_single_pin_rlc()
+
         self.core_components.refresh_components()
         self.logger.info("Deleted {} additional components".format(i))
